requirements.txt
# Bible Clock v4 - Core Dependencies
# Install with: pip install -r requirements.txt

# Core Dependencies (Required)
Pillow>=10.1.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
psutil>=5.9.0
numpy>=1.24.0

# Web Interface
Flask>=3.0.0
Flask-Cors>=4.0.0

# Voice Control (Optional - requires system audio setup)
pyttsx3>=2.98
SpeechRecognition>=3.10.0
PyAudio>=0.2.11
pvporcupine>=3.0.1
piper-tts>=1.2.0
piper-phonemize>=1.1.0
pydub>=0.25.1
soundfile>=0.12.1

# AI Integration (Optional - requires OpenAI API key)
openai>=1.91.0

# Math and ML Dependencies
scipy>=1.11.0
onnxruntime>=1.22.0

# Utility Libraries
coloredlogs>=15.0.1
humanfriendly>=10.0.0
beautifulsoup4>=4.13.0
tqdm>=4.67.0

# Raspberry Pi Hardware (Install only on Pi)
# RPi.GPIO>=0.7.1
# spidev>=3.7

# Note: Hardware-specific packages are commented out for cross-platform compatibility
# For Raspberry Pi deployment, uncomment the hardware packages above